backend_core/requirements.txt
alembic==0.9.5
Baker==1.3
boto3==1.4.7
botocore==1.7.25
certifi==2017.7.27.1
chardet==3.0.4
cheroot==5.1.0
CherryPy==10.1.0
Clay==2.12.7
clay-flask==2.1.5
click==6.7
colorama==0.3.9
construct==2.5.3
Django==1.11.4
django-allauth==0.34.0
django-braces==1.11.0
django-debug-toolbar==1.8
django-filter==1.0.4
django-haystack==2.6.1
django-model-utils==3.0.0
django-star-ratings==0.5.6
django-static-precompiler==1.6
django-storages==1.6.5
django-tinymce==2.6.0
django-tinymce4-lite==1.5.2
djangorestframework==3.6.3
django-widget-tweaks==1.4.1
docutils==0.14
Flask==0.12.2
future==0.16.0
futures==3.1.1
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.9.3
jsmin==2.2.2
Mako==1.0.7
Markdown==2.6.8
MarkupSafe==1.0
Moar==3.0.0
oauthlib==2.0.2
olefile==0.44
pefile==2017.9.3
Pillow==4.2.1
portend==2.1.2
pycrypto==2.6.1
Pygments==2.2.0
PyJWT==1.5.2
PyMySQL==0.7.11
pysolr==3.6.0
python-dateutil==2.6.1
python-editor==1.0.3
python-openid==2.2.5
python-ptrace==0.9.2
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
s3transfer==0.1.11
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
SQLAlchemy==1.1.13
sqlparse==0.2.3
tempora==1.8
urllib3==1.22
Voodoo==1.4.2
Wand==0.4.4
Werkzeug==0.12.2
